added guest, student, coach, admin routes
