+++ wave-dev/requirements.txt
@@ -1,0 +1,1 @@
+h2o-wave==1.1.1
