+++ wave/requirements.txt
@@ -64,6 +64,7 @@ nest-asyncio==1.6.0
 notebook==7.1.1
 notebook_shim==0.2.4
 numpy==1.26.4
+oauthlib==3.2.2
 overrides==7.7.0
 packaging==23.2
 pandas==2.2.0
@@ -94,6 +95,7 @@ QtPy==2.4.1
 readchar==4.0.5
 referencing==0.33.0
 requests==2.31.0
+requests-oauthlib==2.0.0
 rfc3339-validator==0.1.4
 rfc3986-validator==0.1.1
 rpds-py==0.18.0
